[Enhancement] Make StarOSAgent::deleteShards return directly on empty list (#18110)

+++ fe/fe-core/src/main/java/com/starrocks/lake/StarOSAgent.java
@@ -367,6 +367,9 @@ public class StarOSAgent {
     }
 
     public void deleteShards(Set<Long> shardIds) throws DdlException {
+        if (shardIds.isEmpty()) {
+            return;
+        }
         prepare();
         try {
             client.deleteShard(serviceId, shardIds);
